themes: Add reciepe to create all poossible 3rd Party themes at once
Recipe neutrino-3rd-party-themes creates only the
theme packages which are defined inside  DEPENDS-line.
A default package 'neutrino-3rd-party-themes' package will be not
createtd. So it is only senseful to use this if all theme packages
are required and to build at once.  e.g. as Package group inside
variable 'IMAGE_INSTALL += neutrino-3rd-party-themes' if reqiured.