gitpkgv.bbclass: Mod workaround for SRCPV handling based on DISTRO_VERSION
Implement a conditional workaround in gitpkgv.bbclass to address issues
with SRCPV when using newer DISTRO_VERSIONs greater than 4.x. This patch
checks the DISTRO_VERSION and applies a special handling if the version
is less than 5.0, setting SRCPV to a workspace tag if it's currently set
to the dummy value '999', see: meta/classes/externalsrc.bbclass
- https://git.yoctoproject.org/poky/commit/?id=d4af8f0268a824f5a4f098c507a0d16ed29bbc58

This workaround is designed to ensure compatibility and proper behavior
until SRCPV is phased out in future Yocto/OE releases as indicated by upstream commits.

This workaround is temporary. After the phase-out of the dummy SRCPV value
in future versions, there will be a need to remove this workaround and
adjust any recipes that rely on it to ensure continued proper
functionality and to clean up obsolete code.

See:
- https://git.yoctoproject.org/poky/commit/?id=62afa02d01794376efab75623f42e7e08af08526
- https://git.yoctoproject.org/poky/commit/?id=65318019cd8c6db19ae5d4526a0fa2d8c8ef25fa

Additionally, it was crucial to note the potential necessity for a
configurable workspace tag. The default dummy value of '999', being a
three-digit number, may cause issues if packages built from the workspace
originally have versions that are more than three digits long. A package
manager might ignore such ''999'' development packages on target hardware
if a package with a longer version number is already installed. This could
complicate testing development packages.
By allowing for a customizable workspace version, it becomes feasible to
generate higher-digit versions for certain packages, aiding package managers in always installing the n
ewest version as long as PR is not modified.